% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_datasets.R
\docType{data}
\name{CNV}
\alias{ACC.CNV}
\alias{ACC.CNV,BLCA.CNV,BRCA.CNV,CESC.CNV,CHOL.CNV,COAD.CNV,COADREAD.CNV,DLBC.CNV,ESCA.CNV,GBM.CNV,GBMLGG.CNV,HNSC.CNV,KICH.CNV,KIPAN.CNV,KIRC.CNV,KIRP.CNV,LAML.CNV,LGG.CNV,LIHC.CNV,LUAD.CNV,LUSC.CNV,MESO.CNV,OV.CNV,PAAD.CNV,PCPG.CNV,PRAD.CNV,READ.CNV,SARC.CNV,SKCM.CNV,STAD.CNV,TGCT.CNV,THCA.CNV,THYM.CNV,UCEC.CNV,UCS.CNV,UVM.CNV}
\alias{BLCA.CNV}
\alias{BRCA.CNV}
\alias{CESC.CNV}
\alias{CHOL.CNV}
\alias{CNV}
\alias{COAD.CNV}
\alias{COADREAD.CNV}
\alias{DLBC.CNV}
\alias{ESCA.CNV}
\alias{GBM.CNV}
\alias{GBMLGG.CNV}
\alias{HNSC.CNV}
\alias{KICH.CNV}
\alias{KIPAN.CNV}
\alias{KIRC.CNV}
\alias{KIRP.CNV}
\alias{LAML.CNV}
\alias{LGG.CNV}
\alias{LIHC.CNV}
\alias{LUAD.CNV}
\alias{LUSC.CNV}
\alias{MESO.CNV}
\alias{OV.CNV}
\alias{PAAD.CNV}
\alias{PCPG.CNV}
\alias{PRAD.CNV}
\alias{READ.CNV}
\alias{SARC.CNV}
\alias{SKCM.CNV}
\alias{STAD.CNV}
\alias{STES.CNV}
\alias{TGCT.CNV}
\alias{THCA.CNV}
\alias{THYM.CNV}
\alias{UCEC.CNV}
\alias{UCS.CNV}
\alias{UVM.CNV}
\title{CNV datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.CNV

BLCA.CNV

BRCA.CNV

CESC.CNV

CHOL.CNV

COAD.CNV

COADREAD.CNV

DLBC.CNV

ESCA.CNV

GBM.CNV

GBMLGG.CNV

HNSC.CNV

KICH.CNV

KIPAN.CNV

KIRC.CNV

KIRP.CNV

LAML.CNV

LGG.CNV

LIHC.CNV

LUAD.CNV

LUSC.CNV

MESO.CNV

OV.CNV

PAAD.CNV

PCPG.CNV

PRAD.CNV

READ.CNV

SARC.CNV

SKCM.CNV

STAD.CNV

STES.CNV

TGCT.CNV

THCA.CNV

THYM.CNV

UCEC.CNV

UCS.CNV

UVM.CNV
}
\value{
Data frames with CNV data.
}
\description{
Package provides CNV datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the date: \code{2015-11-01}. The process is described here: \url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.data} family can be found here: \code{browseVignettes("RTCGA")}
Data format is explained here https://wiki.nci.nih.gov/display/TCGA/Retrieving+Data+Using+the+Data+Matrix
Converting \pkg{RTCGA.CNV} datasets from \code{data.frames} to Bioconductor classes is explained here \link{convertTCGA}.
}
\details{
\code{browseVignettes("RTCGA")}
}
\keyword{datasets}

