% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnv_gene_scores.R
\name{get_region_cnv_score}
\alias{get.region.cnv.score}
\alias{get_region_cnv_score}
\title{Copy Number Variation (CNV) scores for gene or region}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
get.region.cnv.score(chr = "12", start = 69240000, stop = 69200000)
}
\arguments{
  \item{chr}{A chromosome of the region}

  \item{start}{Start of the region (in bp) on the
  chromosome}

  \item{stop}{Stop of the region (in bp) on the chromosome}
}
\value{
A data frame with cnv scores in the Segment_Mean column.
For each segment you will find a log2 - 1 score (0 means
there is no duplication nor deletion)
}
\description{
The get.region.cnv.score() function returns Segment_Mean
scores for selected region on selected chromosome. By
default it's MDM2 (chr12:69240000-69200000).
}

