% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnv_datasets.R
\docType{data}
\name{cnv}
\alias{ACC.cnv}
\alias{ACC.cnv,BLCA.cnv,BRCA.cnv,CESC.cnv,CHOL.cnv,COAD.cnv,COADREAD.cnv,DLBC.cnv,ESCA.cnv,GBM.cnv,GBMLGG.cnv,HNSC.cnv,KICH.cnv,KIPAN.cnv,KIRC.cnv,KIRP.cnv,LAML.cnv,LGG.cnv,LIHC.cnv,LUAD.cnv,LUSC.cnv,MESO.cnv,OV.cnv,PAAD.cnv,PCPG.cnv,PRAD.cnv,READ.cnv,SARC.cnv,SKCM.cnv,STAD.cnv,TGCT.cnv,THCA.cnv,THYM.cnv,UCEC.cnv,UCS.cnv,UVM.cnv}
\alias{BLCA.cnv}
\alias{BRCA.cnv}
\alias{CESC.cnv}
\alias{CHOL.cnv}
\alias{COAD.cnv}
\alias{COADREAD.cnv}
\alias{DLBC.cnv}
\alias{ESCA.cnv}
\alias{GBM.cnv}
\alias{GBMLGG.cnv}
\alias{HNSC.cnv}
\alias{KICH.cnv}
\alias{KIPAN.cnv}
\alias{KIRC.cnv}
\alias{KIRP.cnv}
\alias{LAML.cnv}
\alias{LGG.cnv}
\alias{LIHC.cnv}
\alias{LUAD.cnv}
\alias{LUSC.cnv}
\alias{MESO.cnv}
\alias{OV.cnv}
\alias{PAAD.cnv}
\alias{PCPG.cnv}
\alias{PRAD.cnv}
\alias{READ.cnv}
\alias{SARC.cnv}
\alias{SKCM.cnv}
\alias{STAD.cnv}
\alias{TGCT.cnv}
\alias{THCA.cnv}
\alias{THYM.cnv}
\alias{UCEC.cnv}
\alias{UCS.cnv}
\alias{UVM.cnv}
\alias{cnv}
\title{cnv datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.cnv

BLCA.cnv

BRCA.cnv

CESC.cnv

CHOL.cnv

COAD.cnv

COADREAD.cnv

DLBC.cnv

ESCA.cnv

GBM.cnv

GBMLGG.cnv

HNSC.cnv

KICH.cnv

KIPAN.cnv

KIRC.cnv

KIRP.cnv

LAML.cnv

LGG.cnv

LIHC.cnv

LUAD.cnv

LUSC.cnv

MESO.cnv

OV.cnv

PAAD.cnv

PCPG.cnv

PRAD.cnv

READ.cnv

SARC.cnv

SKCM.cnv

STAD.cnv

TGCT.cnv

THCA.cnv

THYM.cnv

UCEC.cnv

UCS.cnv

UVM.cnv
}
\description{
Package provides cnv datasets from The Cancer Genome Atlas
Project for all cohorts types from
\url{http://gdac.broadinstitute.org/}. Data were downloaded
using \link{RTCGA-package}. The process is described here:
\url{http://mi2-warsaw.github.io/RTCGA.data/}.
}
\keyword{datasets}

